Explicitely convert Integer to int for Java 1.4 compatibility.
git-svn-id: https://svn.apache.org/repos/asf/lenya/trunk@731491 13f79535-47bb-0310-9956-ffa450edef68

+++ src/modules/sitetree/java/src/org/apache/lenya/cms/site/tree2/SaxTreeBuilder.java
@@ -80,7 +80,7 @@ public class SaxTreeBuilder extends AbstractLogEnabled implements TreeBuilder, S
             throws SAXException {
         try {
             if (localName.equals(ELEM_SITE)) {
-                int revision = Integer.valueOf(attrs.getValue("revision"));
+                int revision = Integer.valueOf(attrs.getValue("revision")).intValue();
                 SiteTreeImpl tree = this.currentNode.getTree();
                 int latestRevision = tree.getRevision(tree.getRepositoryNode());
                 if (revision != latestRevision) {
